--- a/lib/protocol/SqlString.js
+++ b/lib/protocol/SqlString.js
@@ -1,10 +1,12 @@
 var SqlString = exports;
 
-SqlString.escapeId = function (val, forbidQualified) {
+SqlString.escapeId = function escapeId(val, forbidQualified) {
   if (Array.isArray(val)) {
-    return val.map(function(v) {
-      return SqlString.escapeId(v, forbidQualified);
-    }).join(', ');
+    var sql = '';
+    for (var i = 0; i < val.length; i++) {
+      sql += (i ? ', ' : '') + escapeId(val[i], forbidQualified);
+    }
+    return sql;
   }
 
   if (forbidQualified) {
@@ -57,11 +59,18 @@ SqlString.escape = function(val, stringifyObjects, timeZone) {
   return "'"+val+"'";
 };
 
-SqlString.arrayToList = function(array, timeZone) {
-  return array.map(function(v) {
-    if (Array.isArray(v)) return '(' + SqlString.arrayToList(v, timeZone) + ')';
-    return SqlString.escape(v, true, timeZone);
-  }).join(', ');
+SqlString.arrayToList = function arrayToList(array, timeZone) {
+  var sql = '';
+  for (var i = 0; i < array.length; i++) {
+    var v = array[i];
+    if (i) sql += ', ';
+    if (Array.isArray(v)) {
+      sql += '(' + arrayToList(v, timeZone) + ')';
+    } else {
+      sql += SqlString.escape(v, true, timeZone);
+    }
+  }
+  return sql;
 };
 
 SqlString.format = function(sql, values, stringifyObjects, timeZone) {
@@ -127,17 +136,17 @@ SqlString.bufferToString = function bufferToString(buffer) {
 };
 
 SqlString.objectToValues = function(object, timeZone) {
-  var values = [];
+  var sql = '';
   for (var key in object) {
     var value = object[key];
-    if(typeof value === 'function') {
+    if (typeof value === 'function') {
       continue;
     }
 
-    values.push(this.escapeId(key) + ' = ' + SqlString.escape(value, true, timeZone));
+    sql += (sql ? ', ' : '') + SqlString.escapeId(key) + ' = ' + SqlString.escape(value, true, timeZone);
   }
 
-  return values.join(', ');
+  return sql;
 };
 
 function zeroPad(number, length) {