In today's V8, appending to a string is faster than building an array and calling `.join()`.

### Benchmark

`$ npm install benchmark`

```js
'use strict';

var Benchmark = require('benchmark');
var SqlString = require('mysql/lib/protocol/SqlString');

var ids = ['id', 'name'];
var arrays = [['value', 12]];
var object = {
  id: 1,
  name: 'Some Name',
  email: 'somename@email.com',
  parentId: 20,
  notify: true,
};

new Benchmark.Suite()
  .add('escapeId', () => {
    SqlString.escapeId(ids);
  })
  .add('escape arrays', () => {
    SqlString.escape(arrays);
  })
  .add('escape object', () => {
    SqlString.escape(object);
  })
  .on('cycle', function(event) {
    console.log(String(event.target));
  })
  .run();
```

#### Results:

Test | Before (ops/sec) | After (ops/sec) | Speedup
------|-----------------------|--------------------|------------
`escapeId` | 1,260,802 | 1,771,151        | 40%
`escape arrays` | 1,037,337 | 2,384,038 | 130%
`escape object` | 279,876 | 456,599        | 63%